pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-sqlalchemy>=3.1.1",
    "google-adk>=1.19.0",
    "google-genai>=1.52.0",
    "markdown>=3.10",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.6",
    "pypdf>=6.4.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
]
